add desktop files, themes, vimrc and other stuff

+++ mint_setup_script.sh
@@ -11,11 +11,11 @@ Only run this after configuring the following:
 After the the successful execution of this script, the user (me) is advised to:
 1. Add another layout and change the switch shortcut to alt + shift.
 2. Add Flameshot shortcut and disable system PrtSc shortcut.
-3. Change desktop background to ~/Pictures/desktop_background and enable slideshow with random ordering.
+3. Change desktop background to /home/$user/Pictures/desktop_background and enable slideshow with random ordering.
 4. Configure menu (import menu_settings.json file from linux_stuff folder in the applet configurations menu.)
 5. Log into Firefox and sync.
 6. Install discord (flatpak) from software manager.
-7. Add the installed system monitor desklet.
+7. Add the installed system monitor desklet and two installed extensions.
 ' 
 
 
@@ -24,36 +24,45 @@ if [[ $EUID -ne 0 ]]; then
    exit 1
 fi
 
-
-total_ops_count=10
-
+total_ops_count=18
+user=$(logname)
 
 echo "1/$total_ops_count: disabling WI-FI power management."
 sed -i 's/wifi.powersave = 3/wifi.powersave = 2/' /etc/NetworkManager/conf.d/default-wifi-powersave-on.conf
 
 echo "2/$total_ops_count: updating all packages."
-apt update
-apt upgrade 
+apt update -qq
+apt-get upgrade -y | tail -1
 
-echo "2/$total_ops_count: installing git."
-apt install git
+echo "3/$total_ops_count: installing git."
+apt-get install git -y > /dev/null
 echo "Configuring user name and email."
 git config --global user.name "postmodernist1488"
 git config --global user.email elitic.pantheism@gmail.com
 
-echo "3/$total_ops_count: installing Visual Studio Code."
-wget -q --show-progress -O code.deb https://go.microsoft.com/fwlink/?LinkID=760868
-apt install ./code.deb
-rm code.deb
+echo "4/$total_ops_count: installing Visual Studio Code."
+if dpkg -s code &> /dev/null; then
+    echo "Visual Studio Code is already installed."
+else
+    wget -q --show-progress -O code.deb https://go.microsoft.com/fwlink/?LinkID=760868
+    apt-get install ./code.deb -y > /dev/null
+    rm code.deb
+fi
 
-echo "4/$total_ops_count: installing auto-cpufreq."
-git clone https://github.com/AdnanHodzic/auto-cpufreq.git
-cd auto-cpufreq && ./auto-cpufreq-installer
-auto-cpufreq --install
+echo "5/$total_ops_count: installing auto-cpufreq."
+if [ ! -d auto-cpufreq ] && ! auto-cpufreq --help &> /dev/null; then
+    su $user -c "git clone --quiet https://github.com/AdnanHodzic/auto-cpufreq.git"
+fi
 
+if ! auto-cpufreq --help &> /dev/null; then
+cd auto-cpufreq && ./auto-cpufreq-installer --install > /dev/null 2> auto-cpufreq_error.log
+auto-cpufreq --install > /dev/null
+cd ..
+else echo "Auto-cpufreq is already installed."
+fi
 
 charge_threshold=60
-echo "5/$total_ops_count: setting charge threshold to $charge_threshold%."
+echo "6/$total_ops_count: setting charge threshold to $charge_threshold%."
 
 echo "[Unit]
 Description=Set the battery charge threshold
@@ -71,44 +80,74 @@ WantedBy=multi-user.target" > /etc/systemd/system/battery-charge-threshold.servi
 systemctl enable battery-charge-threshold.service 
 systemctl start battery-charge-threshold.service
 
-echo "6/$total_ops_count: installing Spotify."
-apt install spotify-client
+echo "7/$total_ops_count: installing Spotify."
+apt-get install spotify-client -y > /dev/null
 
-echo "7/$total_ops_count: installing Flameshot."
-apt install flameshot
+echo "8/$total_ops_count: installing Flameshot."
+apt-get install flameshot -y > /dev/null
 echo "Openning Flameshot config to enable launch on startup and disable prompt."
-flameshot config 
+su $user -c "flameshot config" 
 
-echo "8/$total_ops_count: copying pictures for desktop background."
-cp ./desktop_background ~/Pictures/
+echo "9/$total_ops_count: copying pictures for desktop background."
+su $user -c "cp -r desktop_background /home/$user/Pictures/" 
+
+echo "10/$total_ops_count: adding grub background picture."
+su $user -c "cp ameer-basheer-gV6taBJuBTk-unsplash.jpg /boot/grub"
+update-grub 2> /dev/null
+
+echo "11/$total_ops_count: installing extensions and desklets."
+
+if [ ! -e /home/$user/.local/share/cinnamon/extensions ]; then
+su $user -c "mkdir -p /home/$user/.local/share/cinnamon/extensions"
+fi
+if [ ! -e /home/$user/.local/share/cinnamon/desklets ]; then
+su $user -c "mkdir -p /home/$user/.local/share/cinnamon/desklets"
+fi
+
+if [ ! -e /home/$user/.local/share/cinnamon/extensions/transparent-panels@germanfr ]; then
+    echo "Installing transparent panels."
+    wget -q --show-progress -O transparent-panels@germanfr.zip https://cinnamon-spices.linuxmint.com/files/extensions/transparent-panels@germanfr.zip?time=1652001340
+    su $user -c "unzip -qq transparent-panels@germanfr.zip -d /home/$user/.local/share/cinnamon/extensions"
+fi
+ 
+if [ ! -e /home/$user/.local/share/cinnamon/extensions/gTile@shuairan ]; then
+    echo "Installing gTile."
+    wget -q --show-progress -O gTile@shuairan.zip https://cinnamon-spices.linuxmint.com/files/extensions/gTile@shuairan.zip?time=1652001340
+    su $user -c "unzip -qq gTile@shuairan.zip -d /home/$user/.local/share/cinnamon/extensions"
+fi
+
+if [ ! -e /home/$user/.local/share/cinnamon/desklets/simple-system-monitor@ariel ]; then
+    echo "Installing simple system monitor desklet."
+    wget -q --show-progress -O simple-system-monitor@ariel.zip https://cinnamon-spices.linuxmint.com/files/desklets/simple-system-monitor@ariel.zip?time=1652002289
+    su $user -c "unzip -qq simple-system-monitor@ariel.zip -d /home/$user/.local/share/cinnamon/desklets"
+fi
+
+#Removing unzipped archives.
+if [ -e *.zip ]; then
+    rm *.zip
+fi
+
+echo "12/$total_ops_count: installing vim and copying .vimrc."
+apt-get install vim -y > /dev/null
+su $user -c "cp .vimrc /home/$user"
+
+echo "13/$total_ops_count: importing settings for gnome terminal."
+su $user -c "dconf load /org/gnome/terminal/ < gnome_terminal_settings_backup.txt"
+
+echo "14/$total_ops_count: installing pyglet."
+su $user -c "pip install pyglet | tail -1" 
+
+echo "15/$total_ops_count: installing Steam."
+apt-get install steam -y > /dev/null
+
+echo "16/$total_ops_count: installing GIMP."
+apt-get install gimp -y > /dev/null
+
+echo "17/$total_ops_count: installing htop."
+apt-get install htop -y > /dev/null
+
+echo "18/$total_ops_count: copying desktop files."
+su $user -c "cp -r Desktop/ /home/$user"
 
 
-echo "9/$total_ops_count: adding grub background picture."
-cp ./ameer-basheer-gV6taBJuBTk-unsplash.jpg /boot/grub
-update-grub
-
-echo "10/$total_ops_count: installing extensions and desklets."
-echo "Installing transparent panels."
-wget -q --show-progress -O transparent-panels@germanfr.zip https://cinnamon-spices.linuxmint.com/files/extensions/transparent-panels@germanfr.zip?time=1652001340
-unzip -qq transparent-panels@germanfr.zip -d ~/.local/share/cinnamon/extensions 
-echo "Installing gTile."
-wget -q --show-progress -O gTile@shuairan.zip https://cinnamon-spices.linuxmint.com/files/extensions/gTile@shuairan.zip?time=1652001340
-unzip -qq gTile@shuairan.zip -d ~/.local/share/cinnamon/extensions
-echo "Installing simple system monitor desklet."
-wget -q --show-progress -O simple-system-monitor@ariel.zip https://cinnamon-spices.linuxmint.com/files/desklets/simple-system-monitor@ariel.zip?time=1652002289
-unzip -qq simple-system-monitor@ariel.zip -d ~/.local/share/cinnamon/desklets
-echo "Removing unzipped archives."
-rm *.zip
-
-echo "11/$total_ops_count: installing vim and adding .vimrc."
-apt install vim
-echo ":set number relativenumber
-:highlight LineNr ctermfg=grey
-:set tabstop=4" > ~/.vimrc
-
-echo "12/$total_ops_count: importing settings for gnome terminal."
-dconf load /org/gnome/terminal/ < gnome_terminal_settings_backup.txt
-
-echo "13/$total_ops_count: installing pyglet."
-pip install pyglet | tail -1
